+++ activerecord/lib/active_record/connection_adapters/abstract_mysql_adapter.rb
@@ -6,6 +6,17 @@ module ActiveRecord
       include Savepoints
 
       class SchemaCreation < AbstractAdapter::SchemaCreation
+        def visit_TableDefinition(o)
+          create_sql = "CREATE#{' TEMPORARY' if o.temporary} TABLE "
+          create_sql << "#{quote_table_name(o.name)} "
+          statements = []
+          statements.concat(o.columns.map { |c| accept c })
+          statements.concat(o.indexes.map { |(column_name, options)| index_in_create(o.name, column_name, options) })
+          create_sql << "(#{statements.join(', ')}) "
+          create_sql << "#{o.options}"
+          create_sql << " AS #{@conn.to_sql(o.as)}" if o.as
+          create_sql
+        end
 
         def visit_AddColumn(o)
           add_column_position!(super, column_options(o))
@@ -28,6 +39,11 @@ module ActiveRecord
             sql << " AFTER #{quote_column_name(options[:after])}"
           end
           sql
+        end
+
+        def index_in_create(table_name, column_name, options)
+          index_name, index_type, index_columns, index_options, index_algorithm, index_using = @conn.send(:add_index_options, table_name, column_name, options)
+          "#{index_type} INDEX #{quote_column_name(index_name)} #{index_using} (#{index_columns})#{index_options} #{index_algorithm}".gsub('  ', ' ').strip
         end
       end
 
@@ -223,6 +239,10 @@ module ActiveRecord
       # http://bugs.mysql.com/bug.php?id=39170
       def supports_transaction_isolation?
         version[0] >= 5
+      end
+
+      def supports_indexes_in_create?
+        true
       end
 
       def native_database_types